pyproject.toml
[tool.poetry]
name = "fuck-venmo"
version = "0.1.0"
description = "Fuck Venmo"
authors = ["Radon Rosborough <radon@intuitiveexplanations.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.3"
lxml = "^5.1.0"
html2text = "^2020.1.16"
gql = {version = "^3.5.0", extras = ["requests"]}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
